任务队列bug修复，增加时间操作函数

+++ comm/timer.h
@@ -14,7 +14,7 @@ class timer {
 public:
 	timer();
 	void start();
-	double get_time();
+	int get_time();
 	virtual ~timer();
 
 private:
